Moving ContextVerificationMetadata out of the fdc3-standard (although it remains documented with a callout)
