+++ package.json
@@ -18,8 +18,8 @@
     "dist"
   ],
   "workspaces": [
-    "packages/fdc3-schema",
     "packages/fdc3-context",
+    "packages/fdc3-schema",
     "packages/fdc3-standard",
     "packages/testing",
     "packages/fdc3-agent-proxy",
