+++ toolbox/fdc3-example-apps/server-apps/security-demo-app1/src/app1.ts
@@ -1,5 +1,6 @@
 import { DesktopAgent, PrivateChannel } from '@finos/fdc3';
 import type { Context } from '@finos/fdc3-context';
+import type { AntiReplayClaims, DetachedSignature } from '@finos/fdc3-standard';
 import { connectRemoteHandlers, type ExchangeDataMessage, type FDC3Handlers } from '@finos/fdc3-security';
 import { createLogEntry } from '../../../common/src/security-demo/logging';
 import {
@@ -109,7 +110,7 @@ async function raiseGetPricesIntent(fdc3: DesktopAgent, remoteHandlers: FDC3Hand
     const signResult = (await remoteHandlers.exchangeData('request-prices', {
       context: instrument,
       intent: 'demo.GetPrices',
-    })) as { signature: unknown; antiReplay: unknown };
+    })) as { signature: DetachedSignature; antiReplay: AntiReplayClaims };
     const { signature, antiReplay } = signResult;
 
     const resolution = await fdc3.raiseIntent('demo.GetPrices', instrument, null, { signature, antiReplay });
